+++ config/keys_dev.js
@@ -1,0 +1,4 @@
+module.exports = {
+  uTellyApi: process.env.UTELLY_API_KEY,
+  omdbApi: process.env.OMDB_API_KEY
+};
