+++ server.js
@@ -1,6 +1,5 @@
 const express = require('express');
 const request = require('request');
-
 const path = require('path')
 
 const app = express();
@@ -10,26 +9,10 @@ app.use(express.json())
 app.use(express.urlencoded( { extended: false }))
 
 
-// app.use(express.static())
-app.get('/getFeaturedMovie', (req, res) => {
 
-  var options = {
-  method: 'GET',
-  url: 'https://utelly-tv-shows-and-movies-availability-v1.p.rapidapi.com/lookup',
-  qs: {term: 'bojack', country: 'us'},
-  headers: {
-    'x-rapidapi-host': 'utelly-tv-shows-and-movies-availability-v1.p.rapidapi.com',
-    'x-rapidapi-key': '12e1e06344mshbbfb57c2be00fdfp16c653jsnc7ee0c0aab1d',
-    useQueryString: true
-  }
-};
-
-    request(options, function (error, response, body) {
-    	 if (!error && response.statusCode == 200) {
-         res.send(body)
-       }
-    })
-  })
+const featuredMovie = require('./routes/api/featured_movie')
+// Get featuredMovie Route
+app.use('/getFeaturedMovie', featuredMovie);
 
 // Set Static Folders
 app.use(express.static(path.join(__dirname, 'client', 'build')))
@@ -40,7 +23,6 @@ app.get("*", function(req, res) {
 // app.get('/', (req, res) => {
 //   res.sendFild(path.join(__dirname, 'client', 'build', 'index.html'))
 // })
-
